[chore]: Change code headers to SPDX License headers (#1145)
* Remove existing license headers

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>

* Remove existing license headers with additional copyrights

Contains copyright into to persons outside of LMCache authors.

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>

* Add SPDX license headers

A Python script is added which checks Python files for the license header
and add it if it is missing, it will add the header. This script is added as a local
pre-commit hook.

The Python script is copied from vLLM repository on GitHub.
For the original code, see [vLLM](https://github.com/vllm-project/vllm/blob/main/tools/check_spdx_header.py)

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>

* Add SPDX license hjeader to C/C++ files

Added manually as script is for Python files only.

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>

---------

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>
Co-authored-by: siddhantray <siddhant.r98@gmail.com>