[Misc]: Improve the description provided for PRs (#1679)
Improve description provided for PRs

This commit updates the PR template so that it may guide users to
provide more details about the PR being submitted. At the moment, it is
usually just issue number that it fixes.

The value is twofold:

1. Its helps the reviewers understand what the PR does and if needed
2. It help others that find the PR later and understand the reason
behind the changes

It also moves details about identifier prefixes to the contributor guide
which is a more relevant place for them.

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>